requirements.txt
fastapi
openai
openai-whisper
git+https://github.com/openai/whisper.git
uvicorn
torch
python-dotenv